+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 * PR #1380: Fixed ImpersonatedCredentials serialization
+* PR #1381: Added the option to set custom credentials scopes
 
 ## 0.42.3 - 2025-05-22
 * PR #1369: Catalog enhancements